COMMERCE-2953 - commerce-product-service - Avoid CI failures as data for tests is being populated directly to test DBs(skipping ExpandoColumnModelListener.onAfterUpdate)

+++ modules/apps/commerce/commerce-product-service/src/main/java/com/liferay/commerce/product/internal/model/listener/ExpandoColumnModelListener.java
@@ -108,6 +108,10 @@ public class ExpandoColumnModelListener
 				_classNameLocalService.getClassNameId(
 					CPDefinitionOptionValueRel.class.getName()));
 
+		if (cpDefinitionOptionValueRelExpandoTable == null) {
+			return;
+		}
+
 		_expandoColumnLocalService.deleteColumn(
 			cpDefinitionOptionValueRelExpandoTable.getTableId(),
 			cpOptionValueExpandoColumn.getName());
